+++ src/module.ts
@@ -13,8 +13,10 @@ import { DiscoveryModule, Reflector } from '@nestjs/core';
 import { GraphStudioOptions, GraphStudioAsyncOptions, mergeOptions } from './options';
 import { GRAPH_STUDIO_OPTIONS } from './constants';
 import { SnapshotCollector } from './snapshot/collector';
+import { GraphQLOperationCollector } from './snapshot/graphql-collector';
 import { GraphStudioController } from './http/controller';
 import { GraphAnalyzer } from './analysis/analyzer';
+import { GraphQlStudioController } from './http/graphql-controller';
 
 @Module({})
 export class GraphStudioModule implements NestModule {
@@ -37,14 +39,15 @@ export class GraphStudioModule implements NestModule {
       Reflector,
       SnapshotCollector,
       GraphAnalyzer,
+      GraphQLOperationCollector,
     ];
 
     return {
       module: GraphStudioModule,
       imports: [DiscoveryModule],
       providers,
-      controllers: [GraphStudioController],
-      exports: [GRAPH_STUDIO_OPTIONS, SnapshotCollector],
+      controllers: [GraphQlStudioController, GraphStudioController],
+      exports: [GRAPH_STUDIO_OPTIONS, SnapshotCollector, GraphQLOperationCollector],
     };
   }
 
@@ -61,13 +64,14 @@ export class GraphStudioModule implements NestModule {
       Reflector,
       SnapshotCollector,
       GraphAnalyzer,
+      GraphQLOperationCollector,
     ];
 
     return {
       module: GraphStudioModule,
       imports: [DiscoveryModule],
       providers,
-      controllers: [GraphStudioController],
+      controllers: [GraphQlStudioController, GraphStudioController],
       exports: [GRAPH_STUDIO_OPTIONS],
     };
   }